+++ Sources/Sentry/include/SentrySwift.h
@@ -1,6 +1,10 @@
 #ifndef SentrySwift_h
 #define SentrySwift_h
 
-@import SentryPrivate;
+#if __has_include("Sentry-Swift.h")
+#    import "Sentry-Swift.h"
+#else
+#    import "Sentry/Sentry-Swift.h"
+#endif
 
 #endif