+++ Sources/Swift/Tools/UrlSanitized.swift
@@ -1,15 +1,15 @@
 import Foundation
 
 @objcMembers
-public class UrlSanitized: NSObject {
-    public static let SENSITIVE_DATA_SUBSTITUTE = "[Filtered]"
+class UrlSanitized: NSObject {
+    static let SENSITIVE_DATA_SUBSTITUTE = "[Filtered]"
     private var components: URLComponents?
 
-    public var query: String? { components?.query }
-    public var queryItems: [URLQueryItem]? { components?.queryItems }
-    public var fragment: String? { components?.fragment }
+    var query: String? { components?.query }
+    var queryItems: [URLQueryItem]? { components?.queryItems }
+    var fragment: String? { components?.fragment }
 
-    public init(URL url: URL) {
+    init(URL url: URL) {
         components = URLComponents(url: url, resolvingAgainstBaseURL: false)
 
         if components?.user != nil {
@@ -21,7 +21,7 @@ public class UrlSanitized: NSObject {
         }
     }
 
-    public var sanitizedUrl: String? {
+    var sanitizedUrl: String? {
         guard var result = self.components?.string else { return nil }
         if let end = result.firstIndex(of: "?") ?? result.firstIndex(of: "#") {
             result = String(result[result.startIndex..<end])